+++ Sources/IdentityClient/Client/ServiceHub.swift
@@ -8,31 +8,23 @@
 import Foundation
 
 
-public final class ServiceHub: @unchecked Sendable {
+public final class ServiceHub: Sendable {
     
-    private let lock        = CriticalSectionLock()
-    private let builderLock = CriticalSectionLock()
+    internal let authRepository: AuthRepository
+    internal let accountRepository: MyAccountRepository
+    internal let devicesRepository: DevicesRepository
+    internal let userRepository: UserInfoRepository
+    internal let thisDeviceRepository: ThisDeviceRepository
     
-    private var authorization: AuthorizationService!
-    private var account: AccountService!
-    private var devices: DevicesService!
-    private var user: UserService!
-    private var userRegistration: UserRegistrationService!
-    
-    private let processorBuilder: () -> RequestProcessor
-    private let configuration: IdentityClient.Configuration
-    private let options: IdentityClient.Options
-    private let errorParser: ErrorParser
-    
-    private let storage: ValueStorage
-    private let secureStorage: SecureStorage
-    private let tokenStorage: TokenStorage
-    private let deviceInfo: CurrentDeviceInfoProvider
-    private let client: Client
-    
+    let processor: RequestProcessorWrapper
+    let authorizationService: AuthorizationService
+    let accountService: AccountService
+    let devicesService: DevicesService
+    let userService: UserService
+    let registrationService: UserRegistrationService
     
     init(
-        processorBuilder: @escaping () -> RequestProcessor,
+        processorBuilder: @escaping @Sendable () -> RequestProcessor,
         configuration: IdentityClient.Configuration,
         options: IdentityClient.Options,
         storage: ValueStorage,
@@ -42,134 +34,54 @@ public final class ServiceHub: @unchecked Sendable {
         client: Client,
         errorParser: ErrorParser,
     ) {
-        self.processorBuilder = processorBuilder
-        self.configuration    = configuration
-        self.options          = options
-        self.storage          = storage
-        self.secureStorage    = secureStorage
-        self.tokenStorage     = tokenStorage
-        self.deviceInfo       = deviceInfo
-        self.client           = client
-        self.errorParser      = errorParser
-    }
-    
-    private var requestProcessor: RequestProcessorWrapper!
-    
-    var processor: RequestProcessorWrapper {
-        builderLock.withLock {
-            if requestProcessor == nil {
-                let wrapped = RequestProcessorWrapper(
-                    processor: processorBuilder(),
-                    tokenAccessor: tokenStorage)
-                
-                requestProcessor = wrapped
-            }
-            
-            return requestProcessor
-        }
-    }
-    
-    internal lazy
-    var authRepository: AuthRepository = {
-        DefaultRepositoryFactory.authRepository(
+        let processor = RequestProcessorWrapper(
+            processor: processorBuilder(),
+            tokenAccessor: tokenStorage)
+        let authRepository = DefaultRepositoryFactory.authRepository(
             configuration: configuration,
             requestProcessor: processor)
-    }()
-    
-    internal lazy
-    var accountRepository: MyAccountRepository = {
-        DefaultRepositoryFactory.myAccountRepository(
+        let accountRepository = DefaultRepositoryFactory.myAccountRepository(
             configuration: configuration,
             requestProcessor: processor)
-    }()
-    
-    internal lazy
-    var devicesRepository: DevicesRepository = {
-        DefaultRepositoryFactory.devicesRepository(
+        let devicesRepository = DefaultRepositoryFactory.devicesRepository(
             configuration: configuration,
             requestProcessor: processor)
-    }()
-    
-    internal lazy
-    var userRepository: UserInfoRepository = {
-        DefaultRepositoryFactory.userRepository(
+        let userRepository = DefaultRepositoryFactory.userRepository(
             configuration: configuration,
             requestProcessor: processor)
-    }()
-    
-    internal lazy
-    var thisDeviceRepository: ThisDeviceRepository = {
-        DefaultRepositoryFactory.thisDeviceRepository(
+        let thisDeviceRepository = DefaultRepositoryFactory.thisDeviceRepository(
             storage: storage,
             secureStorage: secureStorage,
             currentDeviceInfoProvider: deviceInfo)
-    }()
-    
-    
-    var authorizationService: AuthorizationService {
-        lock.withLock {
-            if authorization == nil {
-                self.authorization = .init(
-                    authRepository: authRepository,
-                    accountRepository: accountRepository,
-                    devicesRepository: devicesRepository,
-                    thisDeviceRepository: thisDeviceRepository,
-                    tokenStorage: tokenStorage,
-                    client: client,
-                    configuration: configuration)
-            }
-            
-            return authorization
-        }
-    }
-    
-    var accountService: AccountService {
-        lock.withLock {
-            if account == nil {
-                account = .init(accountRepository: accountRepository)
-            }
-            
-            return account
-        }
-    }
-    
-    var devicesService: DevicesService {
-        lock.withLock {
-            if devices == nil {
-                devices = .init(
-                    identityOptions: options,
-                    serviceProvider: self,
-                    thisDeviceRepository: thisDeviceRepository,
-                    devicesRepository: devicesRepository,
-                    valueStorage: storage,
-                    secureStorage: secureStorage,
-                    errorParser: errorParser,
-                    client: client)
-            }
-            
-            return devices
-        }
-    }
-    
-    var userService: UserService {
-        lock.withLock {
-            if user == nil {
-                user = .init(userRepository: userRepository)
-            }
-            
-            return user
-        }
-    }
-    
-    var registrationService: UserRegistrationService {
-        lock.withLock {
-            if userRegistration == nil {
-                userRegistration = .init(
-                    accountRepository: accountRepository,
-                    errorParser: errorParser)
-            }
-            
-            return userRegistration
-        }
+        let authorizationService = AuthorizationService(
+            authRepository: authRepository,
+            accountRepository: accountRepository,
+            devicesRepository: devicesRepository,
+            thisDeviceRepository: thisDeviceRepository,
+            tokenStorage: tokenStorage,
+            client: client,
+            configuration: configuration)
+        
+        self.processor = processor
+        self.authRepository = authRepository
+        self.accountRepository = accountRepository
+        self.devicesRepository = devicesRepository
+        self.userRepository = userRepository
+        self.thisDeviceRepository = thisDeviceRepository
+        self.authorizationService = authorizationService
+        self.accountService = AccountService(accountRepository: accountRepository)
+        self.devicesService = DevicesService(
+            identityOptions: options,
+            authorizationService: authorizationService,
+            thisDeviceRepository: thisDeviceRepository,
+            devicesRepository: devicesRepository,
+            valueStorage: storage,
+            secureStorage: secureStorage,
+            errorParser: errorParser,
+            client: client)
+        self.userService = UserService(userRepository: userRepository)
+        self.registrationService = UserRegistrationService(
+            accountRepository: accountRepository,
+            errorParser: errorParser)
     }
 }